Tidy `lib/opentelemetry`
