+++ lib/opentelemetry/base.ts
@@ -1,0 +1,3 @@
+export const headers: Record<string, string> = {
+	"signoz-access-token": process.env.SIGNOZ_INGESTION_TOKEN ?? "",
+};
