+++ lib/opentelemetry/index.ts
@@ -1,1 +1,3 @@
-export * from "./setup";
+export * from "./log";
+export * from "./metric";
+export * from "./trace";
