--- lib/opentelemetry/setup.ts
@@ -1,45 +1,0 @@
-import { OTLPLogExporter } from "@opentelemetry/exporter-logs-otlp-http";
-import { OTLPMetricExporter } from "@opentelemetry/exporter-metrics-otlp-http";
-import {
-	BatchLogRecordProcessor,
-	ConsoleLogRecordExporter,
-	LoggerProvider,
-} from "@opentelemetry/sdk-logs";
-import { PeriodicExportingMetricReader } from "@opentelemetry/sdk-metrics";
-import { NoopSpanProcessor } from "@opentelemetry/sdk-trace-base";
-
-const headers: Record<string, string> = {
-	"signoz-access-token": process.env.SIGNOZ_INGESTION_TOKEN ?? "",
-};
-
-// Traces
-
-export const noopSpanProcessor = new NoopSpanProcessor();
-
-// Metrics
-const metricExporter = new OTLPMetricExporter({
-	url: "https://ingest.us.signoz.cloud:443/v1/metrics",
-	headers,
-});
-
-export const metricReader = new PeriodicExportingMetricReader({
-	exporter: metricExporter,
-	exportIntervalMillis: 14000,
-	exportTimeoutMillis: 5000,
-});
-
-// Logs
-const logExporter = new OTLPLogExporter({
-	url: "https://ingest.us.signoz.cloud:443/v1/logs",
-	headers,
-});
-
-export const logRecordProcessor = new BatchLogRecordProcessor(logExporter);
-
-const loggerProvider = new LoggerProvider();
-loggerProvider.addLogRecordProcessor(new BatchLogRecordProcessor(logExporter));
-loggerProvider.addLogRecordProcessor(
-	new BatchLogRecordProcessor(new ConsoleLogRecordExporter()),
-);
-
-export const logger = loggerProvider.getLogger("giselles");
