+++ lib/opentelemetry/trace.ts
@@ -1,0 +1,3 @@
+import { NoopSpanProcessor } from "@opentelemetry/sdk-trace-base";
+
+export const noopSpanProcessor = new NoopSpanProcessor();
